+++ src/tikray/macropipe/__init__.py
@@ -1,6 +1,7 @@
 import tikray.macropipe.lib  # noqa: F401
 
-from .core import MacroPipe, recipe
+from .core import MacroPipe
+from .registry import recipe
 
 __all__ = [
     "recipe",